refactor: Use separate files for encryption
