+++ src/drive/lib/encryption/keys.js
@@ -196,37 +196,3 @@ export const generateAESKey = async ({
     keyUsages || ['encrypt', 'decrypt']
   )
 }
-
-/**
-  Encrypt data with the given key
-
-  * @param {CryptoKey} key      The encryption key
-  * @param {ArrayBuffer} data   The data to encrypt
-  * @param {object} params           Additional parameters
-  * @param {string} algorithm        The key algorithm. Default is "AES-GCM"
-  * @returns {object}
-
-*/
-export const encryptData = async (key, data, { algorithm } = {}) => {
-  const name = algorithm || 'AES-GCM'
-  // The NIST recommands 96 bits iv for AES-GCM: https://web.cs.ucdavis.edu/~rogaway/ocb/gcm.pdf
-  const iv =
-    name === 'AES-GCM'
-      ? window.crypto.getRandomValues(new Uint8Array(16))
-      : null
-  // https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/encrypt
-  const cipher = await crypto.subtle.encrypt({ name, iv }, key, data)
-  return { cipher, iv }
-}
-
-export const decryptData = async (key, data, { algorithm, iv } = {}) => {
-  // https://developer.mozilla.org/en-US/docs/Web/API/SubtleCrypto/decrypt
-  return window.crypto.subtle.decrypt(
-    {
-      name: algorithm || 'AES-GCM',
-      iv: iv
-    },
-    key,
-    data
-  )
-}
